Update to BotBuilder v3.8.3 and other deps as well
