+++ CSharp/Library/QnAMaker/Microsoft.Bot.Builder.CognitiveServices.QnAMaker.nuspec
