+++ CSharp/Library/QnAMaker/Properties/AssemblyInfo.cs
