+++ CSharp/Microsoft.Bot.Builder.CognitiveServices.sln
@@ -6,24 +6,46 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{356A4EE2-684F-461A-94E0-323782163FF8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "QnAMaker", "QnAMaker", "{ACBCF350-DC7A-4F19-A83F-ADB08F5A8778}"
+	ProjectSection(SolutionItems) = preProject
+		Samples\QnAMaker\README.md = Samples\QnAMaker\README.md
+	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Library", "Library", "{F8F3BE81-DC5E-4F9A-ACAA-18C3BF1409BB}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Bot.Builder.CognitiveServices.QnAMaker", "Library\Microsoft.Bot.Builder.CognitiveServices.QnAMaker.csproj", "{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{0F8D5A75-7EB6-45AC-AB20-ED1DD8E3EEAA}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Bot.Builder.CognitiveServices.QnAMaker.Tests", "Test\Microsoft.Bot.Builder.CognitiveServices.QnAMaker.Tests\Microsoft.Bot.Builder.CognitiveServices.QnAMaker.Tests.csproj", "{E7F17CF4-608C-4642-8C7A-00FC166E0760}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleQnABot", "Samples\QnAMaker\SimpleQnABot\SimpleQnABot.csproj", "{1018F341-F0CF-40A6-8CFB-44FB9367C91A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QnABotWithActiveLearning", "Samples\QnAMaker\QnABotWithActiveLearning\QnABotWithActiveLearning.csproj", "{9A68DF07-307D-4658-9882-3D44B35888DE}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QnABotWithOverrides", "Samples\QnAMaker\QnABotWithOverrides\QnABotWithOverrides.csproj", "{37697D9C-679C-43C6-AD6D-BA151C6F3382}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QnABotWithCustomFeedback", "Samples\QnAMaker\QnABotWithCustomFeedback\QnABotWithCustomFeedback.csproj", "{F8F54D13-D043-4603-8BCC-ABA798F65943}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QnAMakerSampleBot", "Samples\QnAMaker\QnAScorableBasedBot\QnAMakerSampleBot.csproj", "{A8BA1066-5695-4D71-ABB4-65E5A5E0C3D4}"
+	ProjectSection(ProjectDependencies) = postProject
+		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{12D94AB2-5FD8-4450-8D42-46FB636A2A7E}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
+EndProject
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,10 +53,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E7F17CF4-608C-4642-8C7A-00FC166E0760}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E7F17CF4-608C-4642-8C7A-00FC166E0760}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E7F17CF4-608C-4642-8C7A-00FC166E0760}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -65,7 +83,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{ACBCF350-DC7A-4F19-A83F-ADB08F5A8778} = {356A4EE2-684F-461A-94E0-323782163FF8}
-		{CBFE0D35-D6C2-4B25-A498-CFC474A2C3A5} = {F8F3BE81-DC5E-4F9A-ACAA-18C3BF1409BB}
 		{E7F17CF4-608C-4642-8C7A-00FC166E0760} = {0F8D5A75-7EB6-45AC-AB20-ED1DD8E3EEAA}
 		{1018F341-F0CF-40A6-8CFB-44FB9367C91A} = {ACBCF350-DC7A-4F19-A83F-ADB08F5A8778}
 		{9A68DF07-307D-4658-9882-3D44B35888DE} = {ACBCF350-DC7A-4F19-A83F-ADB08F5A8778}
